labs/08-ocr/python/read-text/requirements.txt
azure-cognitiveservices-vision-computervision
azure-ai-documentintelligence
pillow
python-dotenv
